[modify]Firmware code integration complete!
