+++ testbench/uart/counter_la_uart.c
@@ -22,6 +22,7 @@
 #include <irq_vex.h>
 #endif
 
+#define USER_PROJ_IRQ0_EN //不確定要不要加這個
 
 extern void uart_write();
 extern void uart_write_char();
@@ -152,6 +153,7 @@ void main()
 
 	//print("\n");
 	//print("Monitor: Test 1 Passed\n\n");	// Makes simulation very long!
+	
 	int* tmp = fir();
 	reg_mprj_datal = *tmp << 16;
 	reg_mprj_datal = *(tmp+1) << 16;
@@ -163,13 +165,7 @@ void main()
 	reg_mprj_datal = *(tmp+7) << 16;
 	reg_mprj_datal = *(tmp+8) << 16;
 	reg_mprj_datal = *(tmp+9) << 16;
-	reg_mprj_datal = *(tmp+10) << 16;
-
-	tmp = matmul();
-	reg_mprj_datal = *tmp << 16;
-	reg_mprj_datal = *(tmp+1) << 16;
-	reg_mprj_datal = *(tmp+2) << 16;
-	reg_mprj_datal = *(tmp+3) << 16;
+	reg_mprj_datal = *(tmp+10) << 16;	
 
 	tmp = qsort();
 	reg_mprj_datal = *tmp << 16;
@@ -181,7 +177,13 @@ void main()
 	reg_mprj_datal = *(tmp+6) << 16;
 	reg_mprj_datal = *(tmp+7) << 16;
 	reg_mprj_datal = *(tmp+8) << 16;
-	reg_mprj_datal = *(tmp+9) << 16;	
+	reg_mprj_datal = *(tmp+9) << 16;
+
+	tmp = matmul();
+	reg_mprj_datal = *tmp << 16;
+	reg_mprj_datal = *(tmp+1) << 16;
+	reg_mprj_datal = *(tmp+2) << 16;
+	reg_mprj_datal = *(tmp+3) << 16;
 
 	reg_mprj_datal = 0xAB510000;
