--- testbench/uart/matmul.h
@@ -1,16 +1,0 @@
-#ifndef _MATMUL_H
-#define _MATMUL_H
-
-#define SIZE 4
-	int AA[SIZE*SIZE] = {0, 1, 2, 3,
-			0, 1, 2, 3,
-			0, 1, 2, 3,
-			0, 1, 2, 3,
-	};
-	int B[SIZE*SIZE] = {1, 2, 3, 4,
-		5, 6, 7, 8,
-		9, 10, 11, 12,
-		13, 14, 15, 16,
-	};
-	int result[SIZE*SIZE];
-#endif
