--- testbench/uart/qsort.c
@@ -1,34 +1,0 @@
-#include "qsort.h"
-
-int __attribute__ ( ( section ( ".mprjram" ) ) ) partition(int low,int hi){
-	int pivot = A[hi];
-	int i = low-1,j;
-	int temp;
-	for(j = low;j<hi;j++){
-		if(A[j] < pivot){
-			i = i+1;
-			temp = A[i];
-			A[i] = A[j];
-			A[j] = temp;
-		}
-	}
-	if(A[hi] < A[i+1]){
-		temp = A[i+1];
-		A[i+1] = A[hi];
-		A[hi] = temp;
-	}
-	return i+1;
-}
-
-void __attribute__ ( ( section ( ".mprjram" ) ) ) sort(int low, int hi){
-	if(low < hi){
-		int p = partition(low, hi);
-		sort(low,p-1);
-		sort(p+1,hi);
-	}
-}
-
-int* __attribute__ ( ( section ( ".mprjram" ) ) ) qsort(){
-	sort(0,SIZE-1);
-	return A;
-}
