--- testbench/uart/qsort.h
@@ -1,6 +1,0 @@
-#ifndef _QSORT_H
-#define _QSORT_H
-
-#define SIZE 10
-int A[SIZE] = {893, 40, 3233, 4267, 2669, 2541, 9073, 6023, 5681, 4622};
-#endif
